backend/azure_functions/requirements.txt
azure-functions==1.20.0
azure-cosmos==4.7.0
PyJWT==2.8.0